- changes about @flags in m_disconnect(). It should be cohenernt with m_connect(), that is unsigned long too. - expanding all ->valid fields to 64bit, as it is exhaused and there is no free bits.

+++ lustre/include/linux/lustre_export.h
@@ -76,7 +76,7 @@ struct obd_export {
         time_t                    exp_last_request_time;
         spinlock_t                exp_lock; /* protects flags int below */
         /* ^ protects exp_outstanding_replies too */
-        int                       exp_flags;
+        unsigned long             exp_flags;
         int                       exp_failed:1,
                                   exp_replay_needed:1,
                                   exp_libclient:1; /* liblustre client? */
